+++ Magento/Mtf/Client/Driver/Selenium/Driver/PageLoaderInterface.php
@@ -6,11 +6,21 @@
 
 namespace Magento\Mtf\Client\Driver\Selenium\Driver;
 
+use Magento\Mtf\Client\Driver\Selenium\RemoteDriver;
+
 /**
  * Interface for load page.
  */
 interface PageLoaderInterface
 {
+    /**
+     * Set driver.
+     *
+     * @param RemoteDriver $driver
+     * @return $this
+     */
+    public function setDriver(RemoteDriver $driver);
+
     /**
      * Wait page to load.
      *